test: DtoFactory의 simpleProfileResponse 메소드를 DummyProfile으로 이동
